+++ src/pages/Home/components/ClothesSection/index.tsx
@@ -1,8 +1,7 @@
-import { getDefaultClothes, getUserLookbookByTemp } from '@/service/clothes';
+import { getDefaultClothes } from '@/service/clothes';
 import { WeatherResponseData, WeatherType } from '@/types/weather';
 import { useQuery } from '@tanstack/react-query';
 import { C, S } from './style';
-import { Outfits } from '@/types/clothes';
 import { ToggleButtonGroup } from '@mui/material';
 import { useState } from 'react';
 import useAppSelector from '@/hooks/useAppSelector';
@@ -10,11 +9,10 @@ import RecommendClothesLoading from './loading';
 import NetworkError from '@/components/NetworkError';
 import { useKeenSlider } from 'keen-slider/react';
 import 'keen-slider/keen-slider.min.css';
-import AddIcon from '@/assets/svg/add.svg?react';
 import { LOOKBOOK_WEATHER_TYPE } from '@/constants/Lookbook/data';
-import { getClothesImageJSX } from '@/utils/clothes';
-import { useNavigate, useSearchParams } from 'react-router-dom';
+import { useSearchParams } from 'react-router-dom';
 import RecommendClothes from './RecommendClothes';
+import LookbookClothes from './LookbookClothes';
 
 const COOL = 'COOL',
   NORMAL = 'NORMAL',
@@ -39,18 +37,18 @@ type ClothesSectionProps = {
 
 const ClothesSection = ({ weather }: ClothesSectionProps) => {
   const geolocation = useAppSelector((state) => state.geolocation.value);
-  const accessToken = useAppSelector((state) => state.auth.accessToken);
   const user = useAppSelector((state) => state.user.info);
   const [searchParams] = useSearchParams();
+
   const tempParamOption = (searchParams.get('option') ??
     NORMAL) as TempCondition;
+  const weatherType = getWeatehrType(weather.extremumTmp);
+
   const [tempCondition, setTempCondition] = useState<TempCondition>(
     Options.has(tempParamOption) ? tempParamOption : NORMAL
   );
   const [currentSlider, setCurrentSlider] = useState(0);
-  const navigate = useNavigate();
 
-  const weatherType = getWeatehrType(weather.extremumTmp);
   const {
     data: recommedClothes,
     isLoading,
@@ -59,13 +57,6 @@ const ClothesSection = ({ weather }: ClothesSectionProps) => {
   } = useQuery({
     queryKey: ['clothes', tempCondition, geolocation?.region, weather],
     queryFn: () => getDefaultClothes({ ...weather, tempCondition }),
-  });
-
-  const { data: lookbook } = useQuery({
-    queryKey: ['user', user?.socialId, 'lookbook', weather, tempCondition],
-    queryFn: () =>
-      getUserLookbookByTemp(weather.extremumTmp, tempCondition, accessToken),
-    enabled: !!user,
   });
 
   const [sliderRef, instanceRef] = useKeenSlider({
@@ -94,21 +85,6 @@ const ClothesSection = ({ weather }: ClothesSectionProps) => {
     setTempCondition(condition);
   };
 
-  const handleLookbookItemClick = (outfit?: Outfits) => () => {
-    let type = Number(weatherType);
-
-    if (tempCondition === 'COOL') type = type - 1;
-    else if (tempCondition === 'WARM') type = type + 1;
-
-    navigate(`/user/lookbook/create?type=${type}`, {
-      state: {
-        outfit,
-        referrer: `/?tab=lookbook&option=${tempCondition}`,
-        tempOption: tempCondition,
-      },
-    });
-  };
-
   if (isError) return <NetworkError handleRefetch={refetch} />;
   return (
     <S.Section>
@@ -127,27 +103,11 @@ const ClothesSection = ({ weather }: ClothesSectionProps) => {
         {user && (
           <>
             <S.SliderItem className='keen-slider__slide'>
-              <S.LookbookList>
-                {lookbook?.map((outfit) => (
-                  <S.LookbookCard
-                    key={outfit.memberOutfitId}
-                    onClick={handleLookbookItemClick(outfit)}
-                  >
-                    {getClothesImageJSX(outfit.topType, outfit.topColor)}
-                    {getClothesImageJSX(outfit.bottomType, outfit.bottomColor)}
-                  </S.LookbookCard>
-                ))}
-
-                {(!lookbook || lookbook.length <= 3) && (
-                  <S.LookbookCard
-                    $content='add'
-                    onClick={handleLookbookItemClick()}
-                  >
-                    <AddIcon />
-                    <span>추가하기</span>
-                  </S.LookbookCard>
-                )}
-              </S.LookbookList>
+              <LookbookClothes
+                weather={weather}
+                weatherType={weatherType}
+                tempCondition={tempCondition}
+              />
             </S.SliderItem>
 
             {currentSlider === 1 && (
